refactor(admin): extraire UserDetailModal d'AdminUsers (G1 décompo)
AdminUsers.vue : 1163 -> 819 lignes. La modale détail utilisateur (markup +
CSS modal-only) devient src/components/admin/UserDetailModal.vue (props user,
emit close, Teleport préservé). Le mapping de rôle dupliqué table+modale est
extrait dans src/utils/userRoles.js (getRoleLabel). Les classes partagées
table<->modale (.role-*/.avatar-*) passent dans src/assets/styles/_admin-badges.scss
(source unique, @use'd par AdminUsers et la modale) : table et modale restent
strictement cohérentes.

Parité par construction : mêmes classes, déclarations CSS déplacées verbatim,
aucun sélecteur composite container+modal. CSS compilé vérifié (tableau garde
role-badge/avatar via partial ; modale a tout son CSS scoped + hovers partagés).

Tests : tests/unit/userRoles.test.js (mapping pur) + tests/unit/UserDetailModal.test.js
(montage : rendu conditionnel, champs, badge/avatar, masquage optionnels, close).
6/6 verts. vite build vert.
